Ajustement servo

+++ code/src/extern_globals.h
@@ -4,7 +4,7 @@ extern int8_t camera_valeurs[128];
 extern uint16_t objectif_vitesse;
 extern int32_t moteur_derniere_erreur;
 extern int32_t moteur_integrale;
-extern uint16_t moteur_compteur;
+extern uint8_t moteur_compteur;
 extern int8_t controle_derniere_erreur;
 extern int16_t controle_integrale;
 extern uint8_t controle_derniere_position;
